+++ up42/version/version_control.py
@@ -19,13 +19,13 @@ def build_outdated_version_message(installed_version, latest_version):
     return f"You're using an outdated version of the UP42 Python SDK: v{installed_version}. A newer version is available: v{latest_version}.\nPlease upgrade to the latest version using **pip install --upgrade up42-py** or conda **conda update -c conda-forge up42-py**."  # pylint: disable=line-too-long # noqa: E501
 
 
-def is_latest_version_check_disabled(get_environment_variable=os.getenv) -> bool:
+def is_latest_version_check_enabled(get_environment_variable=os.getenv) -> bool:
     value = get_environment_variable(ENV_VAR_UP42_DISABLE_VERSION_CHECK)
     if value is None or value.lower() == "false":
-        return False
-    elif value.lower() == "true":
         return True
-    raise ValueError("UP42_DISABLE_VERSION_CHECK must be a bool.")
+    elif value.lower() == "true":
+        return False
+    raise ValueError(f"UP42_DISABLE_VERSION_CHECK must be 'true' or 'false', got '{value}'.")
 
 
 @functools.lru_cache
@@ -33,10 +33,10 @@ def check_is_latest_version(
     installed_version: str,
     warn=warnings.warn,
     build_warning_message=build_outdated_version_message,
-    is_version_check_disabled=is_latest_version_check_disabled,
+    is_version_check_enabled=is_latest_version_check_enabled,
 ):
     try:
-        if not is_version_check_disabled():
+        if is_version_check_enabled():
             latest_version = _get_latest_version()
             if version.Version(installed_version) < latest_version:
                 warn(build_warning_message(installed_version, latest_version))